Revert "Stop referencing repo version of files. Add"
This reverts commit dba79486f03e48aaff0b6b134a2734d45fa94a2a.

Realized it'll be better to have a unique file on each machine
that references the one in the config. This is good because
that way local variables can be set in the local files and then
referenced in the repo files. Also because no need to copy
anything when updating files.